requirements.txt
streamlit
langchain
langchain-groq
validators
python-dotenv
yt-dlp
unstructured
